[추가] TermsOfServiceAgreeButtonView 추가
