+++ SOOUM/SOOUM/Presentations/Onboarding/TermsOfService/OnboardingTermsOfServiceViewController.swift
@@ -15,60 +15,19 @@ import RxSwift
 import SnapKit
 import Then
 
-class TermsOfServiceAgreeButtonView: UIView {
-    let checkImageView = UIImageView().then {
-        $0.image = .termsOfServiceCheck
-        $0.tintColor = .som.gray01
-    }
-    
-    let titleLabel = UILabel().then {
-        $0.typography = .init(
-            fontContainer: BuiltInFont(
-                size: 18,
-                weight: .regular
-            ),
-            lineHeight: 24.5,
-            letterSpacing: 0
-        )
-        $0.text = "약관 전체 동의"
-        $0.textColor = .som.gray01
-    }
-    
-    // MARK: - init
-    convenience init() {
-        self.init(frame: .zero)
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        initConstraint()
-        self.layer.borderColor = UIColor.som.gray01.cgColor
-        self.layer.borderWidth = 1
-        self.layer.cornerRadius = 12
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    // MARK: - initConstraint
-    private func initConstraint() {
-        self.addSubviews(checkImageView)
-        checkImageView.snp.makeConstraints {
-            $0.leading.equalToSuperview().offset(16)
-            $0.top.equalToSuperview().offset(18)
-            $0.centerY.equalToSuperview()
-        }
-        
-        self.addSubviews(titleLabel)
-        titleLabel.snp.makeConstraints {
-            $0.leading.equalTo(checkImageView.snp.trailing).offset(6)
-            $0.centerY.equalToSuperview()
-        }
-    }
-}
-
 class OnboardingTermsOfServiceViewController: BaseNavigationViewController {
+    
+    enum TermsOfService: CaseIterable {
+        case termsOfService
+        case locationService
+        case privacyPolicy
+    }
+    
+    private var agreementStatus: [TermsOfService: Bool] = [
+        .termsOfService: false,
+        .locationService: false,
+        .privacyPolicy: false
+    ]
     
     let guideLabelView = OnboardingGuideLabelView().then {
         $0.titleLabel.text = "숨을 시작하기 위해서는\n약관 동의가 필요해요"
@@ -121,6 +80,30 @@ class OnboardingTermsOfServiceViewController: BaseNavigationViewController {
             $0.trailing.equalToSuperview().offset(-20)
             $0.bottom.equalTo(self.view.safeAreaLayoutGuide).offset(-13)
         }
+    }
+    
+    override func bind() {
+        agreeAllButtonView.rx.tapGesture()
+            .when(.recognized)
+            .subscribe(with: self) { object, _ in
+                let allAgreed = object.agreementStatus.values.allSatisfy { $0 == true }
+                object.updateAllAgreements(to: !allAgreed)
+            }
+            .disposed(by: disposeBag)
+    }
+    
+    /// 선택 상태 전부 업데이트
+    private func updateAllAgreements(to isAgreed: Bool) {
+        TermsOfService.allCases.forEach { agreementStatus[$0] = isAgreed }
+        
+        updateAgreeAllButtonState()
+        
+        termOfServiceTableView.reloadData()
+    }
+    
+    private func updateAgreeAllButtonState() {
+        let allAgreed = agreementStatus.values.allSatisfy { $0 == true }
+        agreeAllButtonView.updateState(isOn: allAgreed)
     }
 }
