+++ SOOUM/SOOUM/Resources/Assets.xcassets/DesignSystem/Images/termsOfServiceCheck.imageset/Contents.json
@@ -17,5 +17,8 @@
   "info" : {
     "author" : "xcode",
     "version" : 1
+  },
+  "properties" : {
+    "template-rendering-intent" : "template"
   }
 }
